+++ ortto_flutter_sdk_platform_interface/lib/src/method_channel_ortto_flutter_sdk.dart
@@ -39,6 +39,11 @@ class MethodChannelOrttoFlutterSdk extends OrttoFlutterSdkPlatformInterface {
   }
 
   @override
+  Future<void> registerDeviceToken(String token) {
+    throw UnimplementedError();
+  }
+
+  @override
   Future<PushPermission> requestPermissions() {
     throw UnimplementedError();
   }