+++ ortto_flutter_sdk/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.2.0
+
+- Prep for publishing
+
+## 0.1.0+3
+
+* Exports UserID and other associated configuration classes
+
+## 0.1.0+2
+
+* Fixes an issue with android kotlin plugin location
+* Updates android library
+
 ## 0.1.0+1
 
 * Prepare for publishing