+++ ortto_flutter_sdk_android/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.1
+- Downgrade required kotlin version
+
+## 0.2.0
+- Prep for publishing
+
+## 0.1.0+2
+
+* Fixes an issue with android kotlin plugin location
+
 ## 0.1.0+1
 
 * Prepare for publishing