+++ ortto_flutter_sdk_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.0
+* Preparation for v1
+
 ## 0.1.0+1
 
 * Prepare for publishing